finalized lab assignments

+++ Lab Works/Assignments/Lab Week 1/task19.java
@@ -1,8 +1,8 @@
 public class task19 {
     public static void main(String[] args) {
-        double a = 8, b = 3, c = 5; // c = sqrt((8/2)^2 + 3^2)
-        double area =  a * b + a * c;
-        double circumference = c * 6;
+        int a = 8, b = 3, c = 5; // c = sqrt((8/2)^2 + 3^2)
+        int area =  a * b + a * c;
+        int circumference = c * 6;
         System.out.println("Area = " + area + "\nCircumference = " + circumference);
     }
 }
